+++ ingest/io/stats_change_processor_test.go
@@ -18,6 +18,12 @@ func TestStatsChangeProcessor(t *testing.T) {
 	}))
 
 	assert.NoError(t, processor.ProcessChange(Change{
+		Type: xdr.LedgerEntryTypeClaimableBalance,
+		Pre:  nil,
+		Post: &xdr.LedgerEntry{},
+	}))
+
+	assert.NoError(t, processor.ProcessChange(Change{
 		Type: xdr.LedgerEntryTypeData,
 		Pre:  nil,
 		Post: &xdr.LedgerEntry{},
@@ -38,6 +44,12 @@ func TestStatsChangeProcessor(t *testing.T) {
 	// Updated
 	assert.NoError(t, processor.ProcessChange(Change{
 		Type: xdr.LedgerEntryTypeAccount,
+		Pre:  &xdr.LedgerEntry{},
+		Post: &xdr.LedgerEntry{},
+	}))
+
+	assert.NoError(t, processor.ProcessChange(Change{
+		Type: xdr.LedgerEntryTypeClaimableBalance,
 		Pre:  &xdr.LedgerEntry{},
 		Post: &xdr.LedgerEntry{},
 	}))
@@ -68,6 +80,12 @@ func TestStatsChangeProcessor(t *testing.T) {
 	}))
 
 	assert.NoError(t, processor.ProcessChange(Change{
+		Type: xdr.LedgerEntryTypeClaimableBalance,
+		Pre:  &xdr.LedgerEntry{},
+		Post: nil,
+	}))
+
+	assert.NoError(t, processor.ProcessChange(Change{
 		Type: xdr.LedgerEntryTypeData,
 		Pre:  &xdr.LedgerEntry{},
 		Post: nil,
@@ -88,16 +106,19 @@ func TestStatsChangeProcessor(t *testing.T) {
 	results := processor.GetResults()
 
 	assert.Equal(t, int64(1), results.AccountsCreated)
+	assert.Equal(t, int64(1), results.ClaimableBalancesCreated)
 	assert.Equal(t, int64(1), results.DataCreated)
 	assert.Equal(t, int64(1), results.OffersCreated)
 	assert.Equal(t, int64(1), results.TrustLinesCreated)
 
 	assert.Equal(t, int64(1), results.AccountsUpdated)
+	assert.Equal(t, int64(1), results.ClaimableBalancesUpdated)
 	assert.Equal(t, int64(1), results.DataUpdated)
 	assert.Equal(t, int64(1), results.OffersUpdated)
 	assert.Equal(t, int64(1), results.TrustLinesUpdated)
 
 	assert.Equal(t, int64(1), results.AccountsRemoved)
+	assert.Equal(t, int64(1), results.ClaimableBalancesRemoved)
 	assert.Equal(t, int64(1), results.DataRemoved)
 	assert.Equal(t, int64(1), results.OffersRemoved)
 	assert.Equal(t, int64(1), results.TrustLinesRemoved)