+++ services/horizon/internal/init.go
@@ -244,6 +244,7 @@ func initIngestMetrics(app *App) {
 	app.prometheusRegistry.MustRegister(app.ingester.Metrics().LedgerIngestionDuration)
 	app.prometheusRegistry.MustRegister(app.ingester.Metrics().StateVerifyDuration)
 	app.prometheusRegistry.MustRegister(app.ingester.Metrics().StateInvalidGauge)
+	app.prometheusRegistry.MustRegister(app.ingester.Metrics().LedgerStatsCounter)
 }
 
 func initTxSubMetrics(app *App) {